refactor: query 로직 변경
